+++ docs/com/checkout/android_sdk/Input/CardInput.html
@@ -111,7 +111,7 @@ var activeTableTab = "activeTableTab";
 <li>android.support.v7.widget.AppCompatEditText</li>
 <li>
 <ul class="inheritance">
-<li>com.checkout.sdk.input.CardInput</li>
+<li>com.checkout.sdk.cardinput.CardInput</li>
 </ul>
 </li>
 </ul>
@@ -369,7 +369,7 @@ extends android.support.v7.widget.AppCompatEditText</pre>
 <div class="block">This method will clear the whitespace in a number string</div>
 </li>
 </ul>
-<a name="setCardListener-com.checkout.sdk.input.CardInput.Listener-">
+<a name="setCardListener-com.checkout.sdk.cardinput.CardInput.Listener-">
 <!--   -->
 </a>
 <ul class="blockListLast">
