+++ sdk/src/main/java/com/checkout/sdk/cardinput/CardInputUseCase.kt
@@ -1,4 +1,4 @@
-package com.checkout.sdk.usecase
+package com.checkout.sdk.cardinput
 
 import android.text.Editable
 import com.checkout.sdk.architecture.UseCase
@@ -28,7 +28,12 @@ open class CardInputUseCase(
         dataStore.cardNumber = sanitized
         dataStore.cvvLength = cardType.maxCvvLength
 
-        return CardInputResult(sanitized, cardType, isCardValid, false)
+        return CardInputResult(
+            sanitized,
+            cardType,
+            isCardValid,
+            false
+        )
     }
 
     private fun sanitizeEntry(entry: String): String {
