+++ sdk/src/main/java/com/checkout/sdk/view/CardDetailsView.java
@@ -18,7 +18,7 @@ import android.widget.TextView;
 import com.android.volley.VolleyError;
 import com.checkout.sdk.CheckoutAPIClient;
 import com.checkout.sdk.input.BillingInput;
-import com.checkout.sdk.input.CardInput;
+import com.checkout.sdk.cardinput.CardInput;
 import com.checkout.sdk.input.CvvInput;
 import com.checkout.sdk.input.DefaultInput;
 import com.checkout.sdk.input.MonthInput;
